PSD# 164041 :In OCR Screen convert Integer (%) to decimal.
git-svn-id: http://192.168.110.32/svn/pennapps/pff/root/branches/PLF-Release-5.10.0@158392 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ pff-model/src/main/java/com/pennant/backend/model/finance/FinOCRDetail.java
@@ -1,5 +1,6 @@
 package com.pennant.backend.model.finance;
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -22,9 +23,9 @@ public class FinOCRDetail extends AbstractWorkflowEntity implements Comparable<F
 	@XmlElement
 	private int stepSequence;
 	@XmlElement
-	private int customerContribution;
+	private BigDecimal customerContribution;
 	@XmlElement
-	private int financerContribution;
+	private BigDecimal financerContribution;
 	private boolean newRecord = false;
 	private FinOCRDetail befImage;
 	private LoggedInUser userDetails;
@@ -85,19 +86,19 @@ public class FinOCRDetail extends AbstractWorkflowEntity implements Comparable<F
 		this.detailID = ocrDetailID;
 	}
 
-	public int getCustomerContribution() {
+	public BigDecimal getCustomerContribution() {
 		return customerContribution;
 	}
 
-	public void setCustomerContribution(int customerContribution) {
+	public void setCustomerContribution(BigDecimal customerContribution) {
 		this.customerContribution = customerContribution;
 	}
 
-	public int getFinancerContribution() {
+	public BigDecimal getFinancerContribution() {
 		return financerContribution;
 	}
 
-	public void setFinancerContribution(int financerContribution) {
+	public void setFinancerContribution(BigDecimal financerContribution) {
 		this.financerContribution = financerContribution;
 	}
